@@ -1,15 +1,36 @@
 {
+	"name": "fuzeworks/core",
+	"description": "FuzeWorks Framework Core",
+	"homepage": "https://techfuze.net/fuzeworks",
+	"license": ["GPL-3.0"],
+	"authors": [
+		{
+			"name": "Abel Hoogeveen",
+			"homepage": "https://myfuze.net"
+		},
+		{
+			"name": "FuzeWorks Community",
+			"homepage": "https://techfuze.net/fuzeworks/contributors"
+		}
+	],
 	"require": {
-    	"php": ">=5.4.0",
+		"php": ">=5.6.0",
 	    "ext-curl": "*",
-	    "ext-json": "*",
+	    "ext-json": "*"
+	},
+	"suggest": {
+	    "smarty/smarty": "Allows using Smarty in templates",
+		"latte/latte": "Allows using Latte in templates"
+	},
+	"require-dev": {
+        "phpunit/phpunit": "5.3.*",
+		"apigen/apigen": "^4.1",
+		"mikey179/vfsStream": "1.1.*",
 	    "smarty/smarty": "~3.1",
 		"tracy/tracy": "*",
 		"latte/latte": "*"
 	},
-    "require-dev": {
-        "phpunit/phpunit": "5.3.*",
-		"apigen/apigen": "^4.1",
-		"mikey179/vfsStream": "1.1.*"
-    }
+	"autoload": {
+		"classmap": ["Core/System/"]
+	}
 }
