@@ -139,6 +139,8 @@ class Logger {
         self::$log_to_file = Config::get('error')->log_to_file;
         self::$logger_template = Config::get('error')->logger_template;
         self::newLevel('Logger Initiated');
+
+        self::loadTracy();
     }
 
     /**
@@ -146,17 +148,17 @@ class Logger {
      * 
      * @return void
      */
-    public static function loadComposer()
+    public static function loadTracy()
     {
         if (class_exists('\Tracy\Debugger', true))
         {
             if (ENVIRONMENT === 'DEVELOPMENT')
             {
-                Debugger::enable(Debugger::DEVELOPMENT, realpath('Application'.DS.'Logs'));
+                Debugger::enable(Debugger::DEVELOPMENT, realpath(Core::$logDir));
             }
             else
             {
-                Debugger::enable(Debugger::PRODUCTION, realpath('Application'.DS.'Logs'));
+                Debugger::enable(Debugger::PRODUCTION, realpath(Core::$logDir));
             }
             self::$useTracy = true;
         }
@@ -279,15 +281,15 @@ class Logger {
 
         Layout::reset();
         Layout::assign('Logs', self::$Logs);
-        Layout::view(self::$logger_template, 'Core'.DS.'Views', true);
+        Layout::view(self::$logger_template, Core::$coreDir . DS . 'Views', true);
     }
 
     public static function logToFile()
     {
         Layout::reset();
         Layout::assign('Logs', self::$Logs);
-        $contents = Layout::get('logger_cli', 'Core'.DS.'Views');
-        $file = 'Application'.DS.'Logs'.DS.'log_latest.php';
+        $contents = Layout::get('logger_cli', Core::$coreDir . DS . 'Views');
+        $file = Core::$logDir .DS. 'Logs'.DS.'log_latest.php';
         if (is_writable($file))
         {
             file_put_contents($file, '<?php ' . $contents);
